+++ R/class_balance.R
@@ -13,12 +13,17 @@ class_balance <- function(object) {
 #' @rdname class_balance
 #' @export
 #' 
-class_balance.nestcv.glmnet <- function(object) {
+class_balance.default <- function(object) {
   ytrain <- unlist(lapply(object$outer_result, '[[', 'ytrain'))
   if (is.numeric(ytrain)) stop("Not classification", call. = FALSE)
   tab <- table(ytrain)
   cat("Training folds:\n")
   print(c(tab))
+  yfinal <- object$yfinal
+  if (!is.null(yfinal)) {
+    cat("Final fit:\n")
+    print(c(table(yfinal)))
+  }
   invisible(tab)
 }
 
@@ -32,13 +37,11 @@ class_balance.nestcv.train <- function(object) {
   tab <- table(ytrain)
   cat("Training folds:\n")
   print(c(tab))
+  yfinal <- object$yfinal
+  if (!is.null(yfinal)) {
+    cat("Final fit:\n")
+    print(c(table(yfinal)))
+  }
   invisible(tab)
 }
 
-
-#' @rdname class_balance
-#' @export
-#' 
-class_balance.outercv <- function(object) {
-  class_balance.nestcv.glmnet(object)
-}